move CarouselDot and CarouselBar into the proper folder
