+++ src/ui/components/CarouselBar/CarouselBar.tsx
